greatly expanded the existing PostNL ACL

+++ app/code/community/TIG/PostNL/Model/Adminhtml/Observer/ShipmentView.php
@@ -41,12 +41,12 @@
 class TIG_PostNL_Model_Adminhtml_Observer_ShipmentView
 {
     /**
-     * The block we want to edit
+     * The block we want to edit.
      */
     const SHIPMENT_VIEW_BLOCK_NAME = 'adminhtml/sales_order_shipment_view';
 
     /**
-     * Observer that adds a print label button to the shipment view page
+     * Observer that adds a print label button to the shipment view page.
      *
      * @param Varien_Event_Observer $observer
      *
@@ -70,7 +70,7 @@ class TIG_PostNL_Model_Adminhtml_Observer_ShipmentView
         /**
          * Checks if the current block is the one we want to edit.
          *
-         * Unfortunately there is no unique event for this block
+         * Unfortunately there is no unique event for this block.
          */
         $block = $observer->getBlock();
         $shipmentViewClass = Mage::getConfig()->getBlockClassName(self::SHIPMENT_VIEW_BLOCK_NAME);
@@ -91,15 +91,23 @@ class TIG_PostNL_Model_Adminhtml_Observer_ShipmentView
             return $this;
         }
 
-        /**
-         * Add a button to print this shipment's shipping labels
-         */
-        $printShippingLabelUrl = $this->getPrintShippingLabelUrl($shipment->getId());
-        $block->addButton('print_shipping_label', array(
-            'label'   => $helper->__('PostNL - Print Shipping Label'),
-            'onclick' => "printLabel('{$printShippingLabelUrl}')",
-            'class'   => 'download',
-        ));
+        $this->addPostnlButtons($block, $shipment);
+
+        return $this;
+    }
+
+    /**
+     * Add new PostNL buttons to the page.
+     *
+     * @param Mage_Adminhtml_BLock_Sales_Order_Shipment_View $block
+     * @param Mage_Sales_Model_Order_Shipment $shipment
+     *
+     * @return $this
+     */
+    public function addPostnlButtons(Mage_Adminhtml_BLock_Sales_Order_Shipment_View $block,
+                                     Mage_Sales_Model_Order_Shipment $shipment)
+    {
+        $helper = Mage::helper('postnl');
 
         /**
          * @var TIG_PostNL_Model_Core_Shipment $postnlShipment
@@ -107,23 +115,45 @@ class TIG_PostNL_Model_Adminhtml_Observer_ShipmentView
         $postnlShipment = Mage::getModel('postnl_core/shipment')->load($shipment->getId(), 'shipment_id');
 
         /**
+         * Check which actions are allowed.
+         */
+        $confirmAllowed           = $helper->checkIsPostnlActionAllowed('confirm');
+        $printAllowed             = $helper->checkIsPostnlActionAllowed('print_label');
+        $deleteLabelsAllowed      = $helper->checkIsPostnlActionAllowed('delete_labels');
+        $resetConfirmAllowed      = $helper->checkIsPostnlActionAllowed(array('reset_confirmation', 'delete_labels'));
+        $sendTrackAndTraceAllowed = $helper->checkIsPostnlActionAllowed('send_track_and_trace');
+
+        /**
+         * Add a button to print this shipment's shipping labels
+         */
+        if ($printAllowed) {
+            $printShippingLabelUrl = $this->getPrintShippingLabelUrl($shipment->getId());
+
+            $block->addButton('print_shipping_label', array(
+                'label'   => $helper->__('PostNL - Print Shipping Label'),
+                'onclick' => "printLabel('{$printShippingLabelUrl}')",
+                'class'   => 'download',
+            ));
+        }
+
+        /**
          * Add a button to reset the shipment's confirmation status
          */
-        if ($postnlShipment->canResetConfirmation()) {
+        if ($postnlShipment->canResetConfirmation() && $resetConfirmAllowed) {
             $resetConfirmationUrl = $this->getResetConfirmationUrl($shipment->getId());
             $resetWarningMessage = $helper->__(
-                'Are you sure that you wish to reset the confirmation status of this shipment? You will need to '
-                . 'confirm this shipment with PostNL again before you can send it. This action will remove all barcodes'
-                . ' and labels associated with this shipment. You can not undo this action.'
+                                          'Are you sure that you wish to reset the confirmation status of this shipment? You will need to '
+                                          . 'confirm this shipment with PostNL again before you can send it. This action will remove all barcodes'
+                                          . ' and labels associated with this shipment. You can not undo this action.'
             );
 
             $block->addButton('reset_confirmation', array(
                 'label'   => $helper->__('PostNL - Change Confirmation'),
                 'onclick' => "deleteConfirm('"
-                                 . $resetWarningMessage
-                                 . "', '"
-                                 . $resetConfirmationUrl
-                                 . "')",
+                    . $resetWarningMessage
+                    . "', '"
+                    . $resetConfirmationUrl
+                    . "')",
                 'class'   => 'delete',
             ));
         }
@@ -131,24 +161,25 @@ class TIG_PostNL_Model_Adminhtml_Observer_ShipmentView
         /**
          * Update the send tracking info button so that it sends our info, instead of the default
          */
-        if ($postnlShipment->isConfirmed()) {
+        if ($postnlShipment->isConfirmed() && $sendTrackAndTraceAllowed) {
             $resendTrackAndTraceUrl = $this->getResendTrackAndTraceUrl($shipment->getId());
+
             $block->updateButton('save', 'label', $helper->__('PostNL - Send Tracking Information'));
             $block->updateButton('save', 'onclick',
-                "deleteConfirm('"
-                    . $helper->__('Are you sure you want to send PostNL tracking information to the customer?')
-                    . "', '" . $resendTrackAndTraceUrl . "')"
+                                 "deleteConfirm('"
+                                 . $helper->__('Are you sure you want to send PostNL tracking information to the customer?')
+                                 . "', '" . $resendTrackAndTraceUrl . "')"
             );
         }
 
         /**
          * Add a button to remove any stored shipping labels for this shipment.
          */
-        if ($postnlShipment->hasLabels() && !$postnlShipment->isConfirmed()) {
+        if ($postnlShipment->hasLabels() && !$postnlShipment->isConfirmed() && $deleteLabelsAllowed) {
             $removeLabelsUrl = $this->getRemoveLabelsUrl($shipment->getId());
             $removeLabelsWarningMessage = $helper->__(
-                "Are you sure that you wish to remove this shipment\'s shipping label? You will need to print a new "
-                . "shipping label before you can send this shipment."
+                                                 "Are you sure that you wish to remove this shipment\'s shipping label? You will need to print a new "
+                                                 . "shipping label before you can send this shipment."
             );
 
             $block->addButton('remove_shipping_labels', array(
@@ -165,7 +196,7 @@ class TIG_PostNL_Model_Adminhtml_Observer_ShipmentView
         /**
          * Add a button to confirm this shipment.
          */
-        if (!$postnlShipment->isConfirmed()) {
+        if (!$postnlShipment->isConfirmed() && $confirmAllowed) {
             $confirmUrl = $this->getConfirmUrl($shipment->getId());
 
             $block->addButton('confirm_shipment', array(
